requirements.txt
flask
torch
transformers
pillow
tqdm
pymilvus
boto3
requests
pathlib